build(deps): bump storybook to 6.5.0-alpha.7, refactor libraries
- [x] update storybook version to 6.5.0-alpha.7;
- [x] refactor libraries, put components into the
components directory;
