+++ .storybook/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "exclude": ["../**/*.spec.ts", "../**/*.mock.ts"],
   "extends": "../tsconfig.base.json",
-  "include": ["../libs/client-core-components/src/lib/**/*.stories.ts", "../libs/client-diagnostics/src/lib/**/*.stories.ts"]
+  "include": [
+    "../libs/client-core-components/src/lib/components/**/*.stories.ts",
+    "../libs/client-diagnostics/src/lib/components/**/*.stories.ts"
+  ]
 }
