+++ libs/client-core-components/src/lib/client-core-components.module.ts
@@ -4,10 +4,10 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { RouterModule } from '@angular/router';
 import { AppClientMaterialModule } from '@app/client-material';
 
-import { AppContentComponent } from './content/content.component';
-import { AppNavbarComponent } from './navbar/navbar.component';
-import { AppThemeToggleComponent } from './theme-toggle/theme-toggle.component';
-import { AppToolbarComponent } from './toolbar/toolbar.component';
+import { AppContentComponent } from './components/content/content.component';
+import { AppNavbarComponent } from './components/navbar/navbar.component';
+import { AppThemeToggleComponent } from './components/theme-toggle/theme-toggle.component';
+import { AppToolbarComponent } from './components/toolbar/toolbar.component';
 
 @NgModule({
   imports: [FlexLayoutModule, CommonModule, AppClientMaterialModule, RouterModule],
