+++ libs/client-core-components/src/lib/components/content/content.component.ts
