+++ libs/client-core-components/src/lib/components/navbar/navbar.component.scss
