+++ libs/client-core-components/src/lib/components/theme-toggle/theme-toggle.component.scss
