+++ libs/client-core-components/src/lib/components/toolbar/toolbar.component.scss
