+++ libs/client-diagnostics/.storybook/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "../tsconfig.json",
   "exclude": ["../**/*.spec.ts"],
-  "include": ["../src/lib/**/*.stories.ts"]
+  "include": ["../src/lib/components/**/*.stories.ts"]
 }
