+++ libs/client-diagnostics/src/lib/client-diagnostics.module.ts
@@ -4,11 +4,11 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { AppClientMaterialModule } from '@app/client-material';
 
 import { AppClientDiagnosticsRoutingModule } from './client-diagnostics-routing.module';
-import { AppDiagnosticsHomeComponent } from './home/diagnostics-home.component';
-import { AppDiagnosticsHomePage } from './home/page/diagnostics-home-page.component';
-import { AppDiagnosticsIndexComponent } from './index/diagnostics-index.component';
-import { AppDiagnosticsInfoComponent } from './info/diagnostics-info.component';
-import { AppDiagnosticsInfoPage } from './info/page/diagnostics-info-page.component';
+import { AppDiagnosticsHomeComponent } from './components/home/diagnostics-home.component';
+import { AppDiagnosticsHomePage } from './components/home/page/diagnostics-home-page.component';
+import { AppDiagnosticsIndexComponent } from './components/index/diagnostics-index.component';
+import { AppDiagnosticsInfoComponent } from './components/info/diagnostics-info.component';
+import { AppDiagnosticsInfoPage } from './components/info/page/diagnostics-info-page.component';
 
 @NgModule({
   imports: [FlexLayoutModule, CommonModule, AppClientMaterialModule, AppClientDiagnosticsRoutingModule],
