+++ libs/client-diagnostics/src/lib/components/home/diagnostics-home.component.ts
