+++ libs/client-diagnostics/src/lib/components/home/page/diagnostics-home-page.component.scss
