+++ libs/client-diagnostics/src/lib/components/index/diagnostics-index.component.ts
