+++ libs/client-diagnostics/src/lib/components/info/diagnostics-info.component.scss
