+++ libs/client-diagnostics/src/lib/components/info/page/diagnostics-info-page.component.scss
